requirements.txt
-i https://pypi.org/simple
click==7.0
flask~=3.0.3
itsdangerous==1.1.0
jinja2==2.10.1
markupsafe==1.1.1
pycrypto==2.6.1
pyyaml==5.1
werkzeug==0.15.2

requests~=2.30.0
SQLAlchemy~=2.0.29